fix($pathToAction): recognize blank string parameters as NaN
In certain route configurations, a parameter may be a blank string (""). The existing code uses
NaN to check if the parameter is numeric, and then parse it as a float. However, `isNaN('')` is
`false`, but `parseFloat('')` is `NaN`, meaning that blank parameters would be converted to
`NaN`. This manifested itself in the browser with some paths being changed to `/my/path NaN`.